+++ src/common/logger.ts
@@ -1,5 +1,0 @@
-import { Logger } from 'halfred-tools';
-
-export const vscLogger = new Logger('vsc');
-
-export const vscliLogger = new Logger('vscli');