+++ src/filter/conditions/getSearchKeyCondition.ts
@@ -1,7 +1,0 @@
-import { Condition } from '../utils';
-
-const getSearchKeyCondition: Condition.Checker = ({ searchKeyReg }) => {
-  return (record) => searchKeyReg.test(record.name);
-};
-
-export default getSearchKeyCondition;