+++ src/input/index.ts
@@ -1,3 +1,0 @@
-export * from './common';
-
-export { parser as inputParser } from './parser';